+++ manim_slides/qt_utils.py
@@ -1,0 +1,14 @@
+"""Qt utils."""
+
+from PySide6.QtWidgets import QApplication
+
+
+def qapp() -> QApplication:
+    """
+    Return a QApplication instance, creating one
+    if needed.
+    """
+    if app := QApplication.instance():
+        return app
+
+    return QApplication([])