+++ tests/test_defaults.py
@@ -1,3 +1,4 @@
+import shutil
 from pathlib import Path
 
 from manim_slides.defaults import CONFIG_PATH, FFMPEG_BIN, FOLDER_PATH
@@ -13,3 +14,9 @@ def test_config_path() -> None:
 
 def test_ffmpeg_bin() -> None:
     assert FFMPEG_BIN == Path("ffmpeg")
+
+
+def test_ffmpeg_bin_exists() -> None:
+    assert (
+        shutil.which(str(FFMPEG_BIN)) is not None
+    ), "If this fails, many other tests will fail"